+++ metadata.rb
@@ -1,13 +1,13 @@
-name             'ibm-openstack-zvm-driver'
+name             'openstack-zvm-driver'
 maintainer       'IBM Corp.'
-maintainer_email 'www.ibm.com'
-license          'All rights reserved'
-description      'Installs/Configures ibm-openstack-zvm-driver'
+maintainer_email 'https://github.com/zVM-Cloud'
+license          'Apache 2.0'
+description      'Installs/Configures openstack-zvm-driver'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '11.1.0'
 
-recipe 'ibm-openstack-zvm-driver::compute', 'Installs Nova driver for z/VM'
-recipe 'ibm-openstack-zvm-driver::network-agent', 'Installs Neutron agent driver for z/VM'
+recipe 'openstack-zvm-driver::compute', 'Installs Nova driver for z/VM'
+recipe 'openstack-zvm-driver::network-agent', 'Installs Neutron agent driver for z/VM'
 
 supports 'redhat'